\name{swfDevice-package}
\Rdversion{1.1}
\alias{swfDevice-package}
\alias{swfDevice}
\docType{package}
\title{
R graphics output in swf (flash) format.  
}
\description{
swfdevice provides the swf() graphics device which allows R graphics to be output as swf files by using the ming library.
}
\details{
\tabular{ll}{
Package: \tab swfDevice\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-10-20\cr
License: \tab GPL (>= 2.0)\cr
LazyLoad: \tab yes\cr
}
Please see the help page for \code{\link{swf}} for more details.
}
\author{
Cameron Bracken \email{cameron.bracken@gmail.com}
}
\references{
The Ming Library: \url{http://www.libming.org/}
}
\keyword{ 
package 
}
\keyword{ 
device
}
\seealso{
\code{\link[swfDevice:swfDevice-package]{swfDevice}}
}
\examples{
#coming soon 
}
